requirements.txt
absl-py==2.2.2
astunparse==1.6.3
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.4.0
git-filter-repo==2.47.0
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
joblib==1.5.0
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
nltk==3.9.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tkcalendar==1.6.1
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.61
zipp==3.21.0